requirements.txt
pyyaml>=6.0
pandas>=1.5.0
openpyxl>=3.1.0
PyPDF2>=3.0.0
python-docx>=0.8.11
openai>=1.0.0
google-generativeai>=0.3.0
